Refactor search classes to extend StarFci for code reuse.
Reorganized various search classes (SpFci, Bfci, GraspFci, Gfci) to extend the newly updated StarFci class. Removed redundant code, such as methods and fields shared across these classes. This change improves maintainability and enforces consistency in search behavior.
